requirements.txt
torch
ftfy
regex
tqdm
Pillow
faiss-cpu
streamlit
git+https://github.com/openai/CLIP.git
